Adds update orcid robot and removes endpoint and job.
closes #5215
