+++ openapi.yml
@@ -197,23 +197,6 @@ paths:
           required: true
           schema:
             $ref: "#/components/schemas/Druid"
-  "/v1/objects/{id}/update_orcid_work":
-    post:
-      tags:
-        - integrations
-      summary: Update Orcid Work
-      description: "Starts a background job to do this, from metadata in the repository"
-      operationId: "objects#update_orcid_work"
-      responses:
-        "202":
-          description: Accepted
-      parameters:
-        - name: id
-          in: path
-          description: ID of object
-          required: true
-          schema:
-            $ref: "#/components/schemas/Druid"
   "/v1/objects/{id}/notify_goobi":
     post:
       tags:
